Add comprehensive integration tests for dbt import feature
- Add 82 integration tests covering dbt manifest parsing, mapper, codegen, and CLI
- Add real dbt project with DuckDB for testing
- Add GitHub Actions workflow for dbt integration tests
- Fix entity column handling: codegen now includes entity columns in schema
  to match mapper behavior (FeatureView.__init__ expects them)
- Fix type mapping: Int32→ValueType.INT32, Float32→ValueType.FLOAT
  (instead of widening to INT64/DOUBLE)
- Add multi-entity column support tests
- Add comprehensive test coverage for edge cases, validation errors,
  special characters, and CLI commands

+++ sdk/python/feast/dbt/mapper.py
@@ -29,9 +29,9 @@ from feast.value_type import ValueType
 # Mapping from FeastType to ValueType for entity value inference
 FEAST_TYPE_TO_VALUE_TYPE: Dict[FeastType, ValueType] = {
     String: ValueType.STRING,
-    Int32: ValueType.INT64,
+    Int32: ValueType.INT32,
     Int64: ValueType.INT64,
-    Float32: ValueType.DOUBLE,
+    Float32: ValueType.FLOAT,
     Float64: ValueType.DOUBLE,
     Bool: ValueType.BOOL,
     Bytes: ValueType.BYTES,
